item: fix overflow of wiki prices

+++ http-service/src/main/java/net/runelite/http/service/item/ItemController.java
@@ -97,7 +97,7 @@ public class ItemController
 	{
 		if (low > 0 && high > 0)
 		{
-			return (low + high) / 2;
+			return (int) (((long) low + high) / 2);
 		}
 		else
 		{
